pyproject.toml
[tool.poetry]
name = "CliCrewAI"
version = "0.1.0"
description = ""
authors = ["Nik Nik.krichko@gmail.com"]

[tool.poetry.dependencies]
python = ">=3.10.0,<=3.12.2"
crewai = "^0.11.0"
unstructured = '==0.10.25'
pyowm = '3.3.0'
tools = "^0.1.9"
python-dotenv = "1.0.0"
wikipedia = "1.4.0"
yfinance = "0.2.35"
sec-api = "1.0.17"
tiktoken = "0.5.2"
faiss-cpu = "1.7.4"
langchain-community = "0.0.10"
langchain-core = "0.1.8"
openai = "1.7.1"
markdown = "3.4.3"
pymarkdownlnt = "0.9.15"
tasks = "^2.8.0"
opencv-python = "4.8.1.78"
google-search-results = "2.4.2"
langchain_openai = "0.0.2"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"